app/requirements.txt
fastapi
uvicorn[standard]
opentelemetry-distro
opentelemetry-instrumentation-fastapi
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
